Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-autofocus-multiple-times-expected.txt

      

FAIL autofocus is run every time a dialog is opened assert_equals: expected Element node <button id="outer-button" autofocus=""></button> but got Element node <input id="input2" autofocus=""></input>
